Move namespace `speed_change` into new header `data_type/speed_change.hpp`
