+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/speed_target_value_type.hpp
@@ -17,6 +17,7 @@
 
 #include <iostream>
 #include <traffic_simulator/data_type/data_types.hpp>
+#include <traffic_simulator/data_type/speed_change.hpp>
 #include <type_traits>
 
 namespace openscenario_interpreter
