+++ simulation/traffic_simulator/src/data_type/data_types.cpp
@@ -16,30 +16,6 @@
 
 namespace traffic_simulator
 {
-
-namespace speed_change
-{
-double RelativeTargetSpeed::getAbsoluteValue(
-  const std::unordered_map<std::string, traffic_simulator_msgs::msg::EntityStatus> & other_status)
-  const
-{
-  if (const auto iter = other_status.find(reference_entity_name); iter == other_status.end()) {
-    THROW_SEMANTIC_ERROR(
-      "Reference entity name ", std::quoted(reference_entity_name),
-      " is invalid. Please check entity ", std::quoted(reference_entity_name),
-      " exists and not a same entity you want to request changing target speed.");
-  } else {
-    switch (type) {
-      default:
-      case Type::DELTA:
-        return iter->second.action_status.twist.linear.x + value;
-      case Type::FACTOR:
-        return iter->second.action_status.twist.linear.x * value;
-    }
-  }
-}
-}  // namespace speed_change
-
 namespace lane_change
 {
 std::ostream & operator<<(std::ostream & stream, const Direction & value)
